Added optional multithreading to dump_frames_linux_64 and fp_project create
This is enabled via the -j flag.

Surprisingly, it actually makes things slower instead of faster.  It appears
somehow the threads are getting serialized.  This may be a result of Python's
oft-malighned Global Interpreter Lock, but I have not confirmed if GIL is the
cause.  I am committing this even though it is counter-productive because it is
optional and also because I may some day look back and see if I can't debug the
serialization.

The editor only uses 1 job when it runs this script since there is no point in
multithreading yet.