Implement 'snap to grid'.
There's a bug with the toggle button not working, the fix will go into the next
commit (where there will be a comment block explaining the situation).  I
could rebase this commit into the next commit, but I'd rather leave this commit
intact as a memorial to the way that the snap_to_grid toggle button was
supposed to be implemented.