Move on_toggle_snap_to_grid_button from freedom_editor.py into level_display.py
The reason for this is that the do_snap_to_grid global variable in
freedom_editor.py was not getting updated by this function when it was in
freedom_editor.py.  I don't know why this was happening, I *know* I had a
"global do_snap_to_grid" line at the top of the function when it was part of
freedom_editor.py

See commit 0ee5f6bb595095f380794694cfed828d8510025e for context.