Get rid of unnecessary sys.stdout.flush calls
These were only there to get data moving through the pipes back when
fp_project.py create and fp_project.py build were being launched as separate
processes by the editor.  Now that they're imported as modules, I don't need
to constantly flush pipes.

If this is ever needed again it can be put in the do_log default implementations
so that it doesn't need to be sprinkled all over the codebase.