Change the "ERROR: unable to find get_internal_image parameter" msg to a WARNING
It's not serious and it only just started to appear with the get_internal_image
changes in the last couple of commits because we're parsing all the
create_foo_int functions instead of just the ones that get called.  The two
functions this message happens to are both for objects that never actually get
instantiated in game (some sort of cut content I guess).